+++ src/qibo/tests/test_models_circuit_fuse.py
